requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.13.0
spotipy==2.23.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.0.0
joblib==1.3.2
requests==2.31.0